+++ examples/accessControlPolicies.py
@@ -7,13 +7,17 @@
 #	This example shows how to create an <accessControlPolicy> resource in an CSE.
 #
 
-import uuid, sys
+import uuid, sys, logging
 sys.path.append('..')
 from onem2mlib import *
 import onem2mlib.constants as CON
 
+loggingLevel = logging.INFO
+
 
 if __name__ == '__main__':
+	logging.basicConfig(level=loggingLevel)
+	logging.getLogger('urllib3').setLevel(logging.CRITICAL)
 
 	# Create session
 	session = Session('http://localhost:8282', 'admin:admin')