+++ examples/contentInstance.py
@@ -8,12 +8,16 @@
 #	Everything is deleted in the end.
 #
 
-import uuid, sys
+import uuid, sys, logging
 sys.path.append('..')
 from onem2mlib import *
 
+loggingLevel = logging.INFO
 
 if __name__ == '__main__':
+	logging.basicConfig(level=loggingLevel)
+	logging.getLogger('urllib3').setLevel(logging.CRITICAL)
+
 	# create session 
 	session = Session('http://localhost:8282', 'admin:admin')