+++ examples/discovery.py
@@ -7,14 +7,18 @@
 #	This example shows how to discover resources.
 #
 
-import uuid, sys
+import uuid, sys, logging
 sys.path.append('..')
 from onem2mlib import *
 import onem2mlib.constants as CON
 from onem2mlib.utilities import *
 
+loggingLevel = logging.INFO
+
 
 if __name__ == '__main__':
+	logging.basicConfig(level=loggingLevel)
+	logging.getLogger('urllib3').setLevel(logging.CRITICAL)
 
 	# Create session & get CSE
 	session = Session('http://localhost:8282', 'admin:admin')