+++ examples/getAllAEs.py
@@ -7,12 +7,16 @@
 #	This example shows how to retrieve a list of all <AE> resources from an CSE.
 #
 
-import uuid, sys
+import uuid, sys, logging
 sys.path.append('..')
 from onem2mlib import *
 
+loggingLevel = logging.INFO
 
 if __name__ == '__main__':
+	logging.basicConfig(level=loggingLevel)
+	logging.getLogger('urllib3').setLevel(logging.CRITICAL)
+
 	# create session 
 	session = Session('http://localhost:8282', 'admin:admin')