+++ examples/remoteCSE.py
@@ -10,11 +10,16 @@
 #	with another CSE..
 #
 
-import sys
+import sys, logging
 sys.path.append('..')
 from onem2mlib import *
 
+loggingLevel = logging.INFO
+
+
 if __name__ == '__main__':
+	logging.basicConfig(level=loggingLevel)
+	logging.getLogger('urllib3').setLevel(logging.CRITICAL)
 
 	# Create session
 	session = Session('http://localhost:8282', 'admin:admin')