+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [IMPROVEMENT] 'find...()' methods are now part of the ResourceBase" class.
 - [FIX] Group resource creation fixed.
 - [IMPROVEMENT] Now using logging for debug output.
+- [IMPROVEMENT] Simplified marshalling.
 
 ## Version 0.7 (2018-05-13)
 - [IMPROVEMENT] Added support for &lt;remoteCSE> resource type.